refactor: axios interceptor
